Greynir SSML
Added support for using `<greynir>` SSML tags, which get translated into
correct SSML syntax depending on the voice engine being used. Created
`speech/norm` folder, which should house all normalization functionality
for TTS.
