+++ queries/arithmetic.py
@@ -37,7 +37,7 @@ import random
 from query import AnswerTuple, ContextDict, Query, QueryStateDict
 from queries import iceformat_float, gen_answer, read_grammar_file
 from tree import Result, Node, TerminalNode
-from queries.util.num import floats_to_text, numbers_to_text
+from speech.norm.num import floats_to_text, numbers_to_text
 
 
 _ARITHMETIC_QTYPE = "Arithmetic"
