+++ queries/userloc.py
@@ -36,7 +36,7 @@ from queries import (
     cap_first,
     read_grammar_file,
 )
-from queries.util.num import numbers_to_text
+from speech.norm.num import numbers_to_text
 from tree import Result, Node
 from iceaddr import iceaddr_lookup, postcodes  # type: ignore
 from geo import (
